+++ neat/neat.go
@@ -1,6 +1,8 @@
+// Package neat implements an map pre-processor for GoldSrc.
 package neat
 
 import (
+	_ "embed"
 	"fmt"
 	"goldutil/goldsrc"
 	"goldutil/goldsrc/qmap"
@@ -10,6 +12,9 @@ import (
 
 	"github.com/google/uuid"
 )
+
+//go:embed goldutil.fgd
+var FGD string
 
 func Neatify(qm *qmap.QMap, mod *os.Root) error {
 	if err := handleMasters(qm); err != nil {